@@ -5,48 +5,63 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --host dev.guardian.theater --port 3000",
-    "lint": "tslint \"src/**/*.ts\"",
+    "start": "ng serve --aot --host dev.guardian.theater --port 3000 --ssl --ssl-key C:\\Users\\cfried\\certificates\\dev.guardian.theater.key --ssl-cert C:\\Users\\cfried\\certificates\\dev.guardian.theater.crt",
+    "lint": "tslint --project src/tsconfig.json \"src/**/*.ts\" --fix",
     "test": "ng test",
+    "prod": "ng build --prod",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
     "e2e": "protractor",
-    "ngh": "ng build --prod && cp dist/index.html dist/404.html && cp ghp/CNAME dist/CNAME && ngh"
+    "ngh": "ng build --prod && cp dist/index.html dist/404.html && cp ghp/CNAME dist/CNAME && ngh",
+    "manifest": "node ./utils/update-definitions.js && node ./utils/parse-definitions.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^4.0.1",
-    "@angular/compiler": "^4.0.1",
-    "@angular/core": "^4.0.1",
-    "@angular/forms": "^4.0.1",
-    "@angular/http": "^4.0.1",
-    "@angular/platform-browser": "^4.0.1",
-    "@angular/platform-browser-dynamic": "^4.0.1",
-    "@angular/router": "^4.0.1",
-    "angular-2-local-storage": "^1.0.1",
-    "angulartics2": "^2.2.1",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^8.2.6",
+    "@angular/common": "^8.2.6",
+    "@angular/compiler": "^8.2.6",
+    "@angular/core": "^8.2.6",
+    "@angular/forms": "^8.2.6",
+    "@angular/platform-browser": "^8.2.6",
+    "@angular/platform-browser-dynamic": "^8.2.6",
+    "@angular/pwa": "^0.803.4",
+    "@angular/router": "^8.2.6",
+    "@angular/service-worker": "^8.2.4",
+    "@fortawesome/angular-fontawesome": "^0.5.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-brands-svg-icons": "^5.11.2",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "angular-2-local-storage": "^3.0.2",
+    "angulartics2": "^8.0.0",
+    "bungie-api-ts": "^2.2.0",
+    "core-js": "^3.2.1",
+    "fast-equals": "^2.0.0",
+    "http-server": "^0.11.1",
+    "idb-keyval": "^3.2.0",
     "intl": "^1.2.5",
-    "ng2-adsense": "^2.1.1",
-    "rxjs": "^5.0.1",
-    "ts-helpers": "^1.1.1",
-    "zone.js": "^0.8.5"
+    "lodash": "^4.17.15",
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.2",
+    "sass": "^1.22.10",
+    "ts-helpers": "^1.1.2",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular/cli": "^1.0.0-rc.4",
-    "@angular/compiler-cli": "^4.0.1",
-    "@types/jasmine": "2.5.53",
-    "@types/node": "^8.0.19",
-    "codelyzer": "~3.1.2",
-    "jasmine-core": "2.7.0",
-    "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.0.2",
+    "@angular-devkit/build-angular": "~0.803.4",
+    "@angular/cli": "^8.3.4",
+    "@angular/compiler-cli": "^8.2.6",
+    "@types/jasmine": "3.4.0",
+    "@types/node": "^12.7.4",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "3.4.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "4.3.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
+    "karma-jasmine": "^2.0.1",
     "karma-remap-istanbul": "^0.6.0",
-    "protractor": "~5.1.1",
-    "ts-node": "3.3.0",
-    "tslint": "^5.1.0",
-    "typescript": "~2.4.2"
+    "protractor": "~5.4.2",
+    "ts-node": "8.3.0",
+    "tslint": "^5.19.0",
+    "typescript": "~3.5.0"
   }
 }
