@@ -1,15 +1,16 @@
 import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
-import { LocalStorageModule } from 'angular-2-local-storage';
-import { AdsenseModule } from 'ng2-adsense';
-import { Angulartics2Module, Angulartics2GoogleAnalytics } from 'angulartics2';
+import { HttpClientModule } from '@angular/common/http';
+import { Angulartics2Module } from 'angulartics2';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 
 import { RoutesModule } from './routes/routes.module';
 
 import { BungieHttpService } from './services/bungie-http.service';
 import { TwitchService } from './services/twitch.service';
+import { MixerService } from './services/mixer.service';
 import { XboxService } from './services/xbox.service';
 import { SettingsService } from './services/settings.service';
 
@@ -19,11 +20,21 @@ import { FrontPageComponent } from './front-page/front-page.component';
 import { AboutComponent } from './about/about.component';
 import { SettingsComponent } from './settings/settings.component';
 import { GuardianComponent } from './guardian/guardian.component';
-import { ActivityComponent } from './activity/activity.component';
+import {
+  ActivityComponent,
+  PgcrEntryComponent
+} from './activity/activity.component';
 import { DestinyHashPipe } from './pipes/destiny-hash.pipe';
 import { TwitchStampPipe } from './pipes/twitch-stamp.pipe';
 import { GtBadgePipe } from './pipes/gt-badge.pipe';
 import { TakeMyMoneyComponent } from './take-my-money/take-my-money.component';
+import { SearchComponent } from './search/search.component';
+import { BungieStatusComponent } from './bungie-status/bungie-status.component';
+import { PlatformAbbrPipe } from './pipes/platform-abbr.pipe';
+import { ServiceWorkerModule } from '@angular/service-worker';
+import { environment } from '../environments/environment';
+import { ManifestService } from './services/manifest.service';
+import { AuthComponent } from './auth/auth.component';
 
 @NgModule({
   declarations: [
@@ -36,30 +47,35 @@ import { TakeMyMoneyComponent } from './take-my-money/take-my-money.component';
     TwitchStampPipe,
     GtBadgePipe,
     ActivityComponent,
+    PgcrEntryComponent,
     DestinyHashPipe,
-    TakeMyMoneyComponent
+    TakeMyMoneyComponent,
+    SearchComponent,
+    BungieStatusComponent,
+    PlatformAbbrPipe,
+    AuthComponent
   ],
   imports: [
     BrowserModule,
+    BrowserAnimationsModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     RoutesModule,
-    LocalStorageModule.withConfig({
-      prefix: 'gt',
-      storageType: 'localStorage'
+    Angulartics2Module.forRoot(),
+    ServiceWorkerModule.register('ngsw-worker.js', {
+      enabled: environment.production,
+      registrationStrategy: 'registerWithDelay:5000'
     }),
-    AdsenseModule.forRoot({
-      adClient: 'ca-pub-7822250090731539',
-      adSlot: 9694174006
-    }),
-    Angulartics2Module.forRoot([Angulartics2GoogleAnalytics])
+    FontAwesomeModule
   ],
   providers: [
     BungieHttpService,
     TwitchService,
+    MixerService,
     XboxService,
-    SettingsService
+    SettingsService,
+    ManifestService
   ],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
