@@ -1,7 +1,8 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Location } from '@angular/common';
 import { SettingsService } from '../services/settings.service';
-import { Subscription } from 'rxjs/Rx';
+import { Subscription } from 'rxjs';
+import { gt } from '../gt.typings';
 
 @Component({
   selector: 'app-settings',
@@ -10,23 +11,28 @@ import { Subscription } from 'rxjs/Rx';
 })
 export class SettingsComponent implements OnInit, OnDestroy {
   private _subLinks: Subscription;
+  private _subDark: Subscription;
 
   public links: gt.Links;
+  public dark: boolean;
 
   constructor(
     public settingsService: SettingsService,
     private location: Location
-  ) { }
+  ) {}
 
   ngOnInit() {
-    this._subLinks = this.settingsService.links
-      .subscribe(links => {
-        this.links = links;
-      });
+    this._subLinks = this.settingsService.links.subscribe(links => {
+      this.links = links;
+    });
+    this._subDark = this.settingsService.dark.subscribe(
+      dark => (this.dark = dark)
+    );
   }
 
   ngOnDestroy() {
     this._subLinks.unsubscribe();
+    this._subDark.unsubscribe();
   }
 
   toggleActivityLink(link: string) {
@@ -41,6 +47,10 @@ export class SettingsComponent implements OnInit, OnDestroy {
     this.settingsService.toggleXboxLink = link;
   }
 
+  toggleDark() {
+    this.settingsService.toggleDark();
+  }
+
   setLanguage(language: string) {
     this.settingsService.setLanguage = language;
   }
@@ -48,5 +58,4 @@ export class SettingsComponent implements OnInit, OnDestroy {
   back() {
     this.location.back();
   }
-
 }
