@@ -1,18 +1,37 @@
-import { Component, OnInit, OnDestroy, Input } from '@angular/core';
-import { ActivityService } from '../services/activity.service';
-import { BungieHttpService } from '../services/bungie-http.service';
-import { TwitchService } from '../services/twitch.service';
-import { XboxService } from '../services/xbox.service';
-import { SettingsService } from '../services/settings.service';
-import { Subscription, Observable } from 'rxjs/Rx';
+import { animate, style, transition, trigger } from '@angular/animations';
+import {
+  Component,
+  HostListener,
+  Input,
+  OnDestroy,
+  OnInit
+} from '@angular/core';
 import { Router } from '@angular/router';
+import { Subscription, combineLatest as observableCombineLatest } from 'rxjs';
+import { gt } from '../gt.typings';
+import { ActivityService } from '../services/activity.service';
+import { GuardianService } from '../services/guardian.service';
+import { SettingsService } from '../services/settings.service';
+import { TwitchService } from '../services/twitch.service';
+import { MixerService } from '../services/mixer.service';
+import { XboxService } from '../services/xbox.service';
 
 @Component({
   selector: 'app-activity',
   templateUrl: './activity.component.html',
   styleUrls: ['./activity.component.scss'],
-  providers: [
-    ActivityService
+  providers: [ActivityService, GuardianService],
+  animations: [
+    trigger('growInShrinkOut', [
+      transition('void => *', [
+        style({ height: 0 }),
+        animate('600ms ease-out', style({ height: '*' }))
+      ]),
+      transition('* => void', [
+        style({ height: '*' }),
+        animate('300ms ease-in', style({ height: 0 }))
+      ])
+    ])
   ]
 })
 export class ActivityComponent implements OnInit, OnDestroy {
@@ -21,25 +40,42 @@ export class ActivityComponent implements OnInit, OnDestroy {
   private subs: Subscription[];
 
   public pgcr: gt.PostGameCarnageReport;
-  public membershipType: number;
+  public clips: gt.Clip[];
+  public filteredClips: gt.Clip[];
+  public links: gt.Links;
+  public mini: boolean;
+  public animationState;
+  public innerWidth: number;
+
+  @HostListener('window:resize', ['$event'])
+  onResize(event) {
+    window.innerWidth < 640 ? (this.mini = true) : (this.mini = false);
+  }
 
   constructor(
     private activityService: ActivityService,
-    private bHttp: BungieHttpService,
     public twitchService: TwitchService,
+    public mixerService: MixerService,
     private xboxService: XboxService,
     private router: Router,
     public settingsService: SettingsService
-  ) { }
+  ) {}
 
   ngOnInit() {
+    window.innerWidth < 640 ? (this.mini = true) : (this.mini = false);
+
+    this.filteredClips = [];
+    this.clips = [];
+    this.links = { guardian: {}, activity: {}, xbox: {} };
+    this.animationState = 'in';
+
     this.activityService.activity = this.activity;
     this.subs = [];
-    this.subs.push(this.activityService.membershipType
-      .subscribe(membershipType => this.membershipType = membershipType)
+    this.subs.push(
+      this.settingsService.links.subscribe(links => (this.links = links))
     );
-    this.subs.push(this.activityService.pgcr
-      .subscribe((pgcr: gt.PostGameCarnageReport) => {
+    this.subs.push(
+      this.activityService.pgcr.subscribe((pgcr: gt.PostGameCarnageReport) => {
         this.pgcr = pgcr;
         if (pgcr && !pgcr.loading) {
           pgcr.loading = {
@@ -50,74 +86,165 @@ export class ActivityComponent implements OnInit, OnDestroy {
           };
         }
         if (this.pgcr) {
+          this.subs.push(
+            this.pgcr.filteredClips$.subscribe(
+              clips => (this.filteredClips = clips)
+            )
+          );
+          this.subs.push(
+            this.pgcr.clips$.subscribe(clips => (this.clips = clips))
+          );
           pgcr.showClips = true;
           let loadingArray = [];
           this.pgcr.entries.forEach(entry => {
-            if (entry.player.bungieNetUserInfo && this.twitchService.twitch[entry.player.bungieNetUserInfo.membershipId]) {
-              loadingArray.push(this.twitchService.twitch[entry.player.bungieNetUserInfo.membershipId]);
+            if (
+              entry.player.bungieNetUserInfo &&
+              this.twitchService.twitch[
+                entry.player.bungieNetUserInfo.membershipId
+              ]
+            ) {
+              loadingArray.push(
+                this.twitchService.twitch[
+                  entry.player.bungieNetUserInfo.membershipId
+                ]
+              );
               this.subs.push(
-                this.twitchService.twitch[entry.player.bungieNetUserInfo.membershipId]
-                  .subscribe(twitch => entry.twitch = twitch)
+                this.twitchService.twitch[
+                  entry.player.bungieNetUserInfo.membershipId
+                ].subscribe(twitch => (entry.twitch = twitch))
+              );
+            } else if (
+              this.twitchService.twitch[
+                entry.player.destinyUserInfo.membershipId
+              ]
+            ) {
+              loadingArray.push(
+                this.twitchService.twitch[
+                  entry.player.destinyUserInfo.membershipId
+                ]
+              );
+              this.subs.push(
+                this.twitchService.twitch[
+                  entry.player.destinyUserInfo.membershipId
+                ].subscribe(twitch => (entry.twitch = twitch))
               );
             }
-            if (this.membershipType === 1 && this.xboxService.xbox[entry.player.destinyUserInfo.displayName]) {
-              loadingArray.push(this.xboxService.xbox[entry.player.destinyUserInfo.displayName]);
-              this.subs.push(
-                this.xboxService.xbox[entry.player.destinyUserInfo.displayName]
-                  .subscribe(xbox => entry.xbox = xbox)
+
+            if (
+              entry.player.bungieNetUserInfo &&
+              this.mixerService.mixer[
+                entry.player.bungieNetUserInfo.membershipId
+              ]
+            ) {
+              loadingArray.push(
+                this.mixerService.mixer[
+                  entry.player.bungieNetUserInfo.membershipId
+                ]
               );
+              this.subs.push(
+                this.mixerService.mixer[
+                  entry.player.bungieNetUserInfo.membershipId
+                ].subscribe(mixer => (entry.mixer = mixer))
+              );
+            } else if (
+              this.mixerService.mixer[entry.player.destinyUserInfo.membershipId]
+            ) {
+              loadingArray.push(
+                this.mixerService.mixer[
+                  entry.player.destinyUserInfo.membershipId
+                ]
+              );
+              this.subs.push(
+                this.mixerService.mixer[
+                  entry.player.destinyUserInfo.membershipId
+                ].subscribe(mixer => (entry.mixer = mixer))
+              );
+            }
+
+            if (
+              entry.player.destinyUserInfo.membershipType === 1 &&
+              this.xboxService.xbox[entry.player.destinyUserInfo.displayName]
+            ) {
+              loadingArray.push(
+                this.xboxService.xbox[entry.player.destinyUserInfo.displayName]
+              );
+              this.subs.push(
+                this.xboxService.xbox[
+                  entry.player.destinyUserInfo.displayName
+                ].subscribe(xbox => (entry.xbox = xbox))
+              );
+            }
+            if (
+              entry.player.destinyUserInfo.membershipType === 4 &&
+              this.xboxService.xboxPC[entry.player.destinyUserInfo.displayName]
+            ) {
+              loadingArray.push(
+                this.xboxService.xboxPC[
+                  entry.player.destinyUserInfo.displayName
+                ]
+              );
+              this.subs.push(
+                this.xboxService.xboxPC[
+                  entry.player.destinyUserInfo.displayName
+                ].subscribe(xbox => (entry.xbox = xbox))
+              );
+            }
+            switch (entry.player.destinyUserInfo.membershipType) {
+              case 1:
+                entry.trn = 'xbl';
+                break;
+              case 2:
+                entry.trn = 'psn';
+                break;
+              case 4:
+                entry.trn = 'pc';
+                break;
             }
           });
-          if (pgcr.activityDetails.mode === 14) {
-            this.pgcr.teams.forEach(team => {
-              team.trialsLink = 'https://trials.report/' + (team.entries[0].player.destinyUserInfo.membershipType === 1 ? 'xbox' : 'ps');
-              team.entries.forEach(entry => {
-                team.trialsLink += '/' + entry.player.destinyUserInfo.displayName;
-              });
-            });
-          }
           this.subs.push(
-            Observable.combineLatest(loadingArray)
-              .subscribe(
-                array => {
-                  let loading = {
-                    message: '',
-                    bungie: false,
-                    xbox: false,
-                    twitch: false
-                  };
-                  array.some(function(item) {
-                    if (item.bungieId && !item.checkedId) {
-                      loading.message = 'Fetching Twitch username for ' + item.displayName + '...';
-                      return true;
-                    }
-                    if (item.bungieId && !item.checkedResponse) {
-                      loading.message = 'Fetching Twitch videos for ' + item.displayName + '...';
-                      return true;
-                    }
-                    if (item.gamertag && !item.checked) {
-                      loading.message = 'Fetching Xbox clips for ' + item.gamertag + '...';
-                      return true;
-                    }
-                  });
-                  array.some(function(item) {
-                    if (item.bungieId && !item.checkedId) {
-                      return loading.bungie = true;
-                    }
-                  });
-                  array.some(function(item) {
-                    if (item.bungieId && !item.checkedResponse) {
-                      return loading.twitch = true;
-                    }
-                  });
-                  array.some(function(item) {
-                    if (item.gamertag && !item.checked) {
-                      return loading.xbox = true;
-                    }
-                  });
-                  pgcr.loading = loading;
+            observableCombineLatest(loadingArray).subscribe(array => {
+              let loading = {
+                message: '',
+                bungie: false,
+                xbox: false,
+                xboxPC: false,
+                twitch: false,
+                mixer: false
+              };
+              array.some(function(item: { bungieId, checkedId, displayName, checkedResponse, checked, gamertag}) {
+                if (item.bungieId && !item.checkedId) {
+                  loading.message =
+                    'Fetching Twitch username for ' + item.displayName + '...';
+                  return true;
                 }
-              )
+                if (item.bungieId && !item.checkedResponse) {
+                  loading.message =
+                    'Fetching Twitch videos for ' + item.displayName + '...';
+                  return true;
+                }
+                if (item.gamertag && !item.checked) {
+                  loading.message =
+                    'Fetching Xbox clips for ' + item.gamertag + '...';
+                  return true;
+                }
+              });
+              array.some(function(item: { bungieId, checkedId }) {
+                if (item.bungieId && !item.checkedId) {
+                  return (loading.bungie = true);
+                }
+              });
+              array.some(function(item: { bungieId, checkedResponse}) {
+                if (item.bungieId && !item.checkedResponse) {
+                  return (loading.twitch = true);
+                }
+              });
+              array.some(function(item: {gamertag, checked}) {
+                if (item.gamertag && !item.checked) {
+                  return (loading.xbox = true);
+                }
+              });
+              pgcr.loading = loading;
+            })
           );
         }
       })
@@ -125,18 +252,21 @@ export class ActivityComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
+    this.animationState = '';
     this.subs.forEach(sub => sub.unsubscribe());
   }
 
   toActivity(activityId) {
-    this.router.navigate([
-      '/activity', this.membershipType, activityId
-    ]);
+    this.router.navigate(['/activity', activityId]);
   }
 
-  toGuardian(displayName, characterId) {
+  toGuardian(membershipType, membershipId) {
     this.router.navigate([
-      '/guardian', displayName, this.membershipType, characterId
+      '/guardian',
+      membershipType,
+      membershipId,
+      'None',
+      0
     ]);
   }
 
@@ -147,6 +277,37 @@ export class ActivityComponent implements OnInit, OnDestroy {
       pgcr.activityDetails.instanceId,
       clip.entry.characterId,
       clip.video.gameClipId
+    ]);
+  }
+
+  stopPropagation(event) {
+    event.stopPropagation();
+  }
+
+  route(route: any[]) {
+    this.router.navigate(route);
+  }
+}
+
+@Component({
+  selector: 'app-pgcr-entry',
+  templateUrl: './pgcr-entry.component.html',
+  styleUrls: ['./activity.component.scss']
+})
+export class PgcrEntryComponent {
+  @Input() pgcr: gt.PostGameCarnageReport;
+  @Input() entry: gt.Entry;
+  @Input() links: gt.Links;
+
+  constructor(private router: Router) {}
+
+  toGuardian(membershipType, membershipId) {
+    this.router.navigate([
+      '/guardian',
+      membershipType,
+      membershipId,
+      'None',
+      0
     ]);
   }
 
