@@ -1,5 +1,6 @@
-import { Angulartics2GoogleAnalytics } from 'angulartics2';
-import { Component } from '@angular/core';
+import { Angulartics2GoogleAnalytics } from 'angulartics2/ga';
+import { Component, Renderer2 } from '@angular/core';
+import { SettingsService } from './services/settings.service';
 
 @Component({
   selector: 'app-root',
@@ -7,5 +8,16 @@ import { Component } from '@angular/core';
   styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-  constructor(angulartics2GoogleAnalytics: Angulartics2GoogleAnalytics) {}
+  constructor(
+    angulartics2GoogleAnalytics: Angulartics2GoogleAnalytics,
+    private settingsService: SettingsService,
+    private renderer: Renderer2
+  ) {
+    angulartics2GoogleAnalytics.startTracking();
+    this.settingsService.dark.subscribe(dark => {
+      dark
+        ? this.renderer.addClass(document.body, 'dark')
+        : this.renderer.removeClass(document.body, 'dark');
+    });
+  }
 }
