@@ -1,772 +1,1772 @@
 export const DestinyActivityModeDefinition = {
   de: {
-    23: {
-      modeName: "Eliminierung"
-    },
-    25: {
-      modeName: "Hexenkessel"
-    },
-    6: {
-      modeName: "Patrouille"
-    },
-    3: {
-      modeName: "Normale Strikes"
-    },
-    17: {
-      modeName: "Heroische Strikes"
-    },
-    19: {
-      modeName: "Eisenbanner"
-    },
-    16: {
-      modeName: "Dämmerung-Strikes"
+    56: {
+      name: 'Privatmatches: Hexenkessel'
     },
     32: {
-      modeName: "Privatmatches"
+      name: 'Privatmatches'
     },
-    4: {
-      modeName: "Raid"
+    54: {
+      name: 'Privatmatches: Countdown'
     },
-    8: {
-      modeName: "Einleitung"
+    48: {
+      name: 'Rumble'
     },
-    12: {
-      modeName: "Konflikt"
+    58: {
+      name: 'Heroisches Abenteuer'
     },
-    22: {
-      modeName: "Arena-Herausforderung"
+    42: {
+      name: 'Prüfungen der Neun: Überleben'
     },
-    24: {
-      modeName: "Rift"
+    47: {
+      name: 'Score-Prestige-Dämmerung'
     },
-    36: {
-      modeName: "Hexenkessel: Vorherrschaft"
+    46: {
+      name: 'Score-Dämmerungs-Strikes'
     },
-    27: {
-      modeName: "Hexenkessel: Rumble"
+    51: {
+      name: 'Privatmatches: Konflikt'
     },
     0: {
-      modeName: "Alle"
-    },
-    30: {
-      modeName: "Herausforderung der Ältesten"
-    },
-    18: {
-      modeName: "Strikes"
-    },
-    21: {
-      modeName: "Arena"
-    },
-    28: {
-      modeName: "Zonenkontrolle"
-    },
-    9: {
-      modeName: "Gefecht"
-    },
-    29: {
-      modeName: "Sparrow-Rennen"
-    },
-    11: {
-      modeName: "Bergung"
-    },
-    20: {
-      modeName: "Gefängnis der Alten"
-    },
-    34: {
-      modeName: "Vorherrschaft"
-    },
-    13: {
-      modeName: "Rumble"
+      name: 'Alle'
     },
     31: {
-      modeName: "Vorherrschaft"
-    },
-    7: {
-      modeName: "PVE"
-    },
-    26: {
-      modeName: "Hexenkessel: Konflikt"
-    },
-    2: {
-      modeName: "Story"
-    },
-    14: {
-      modeName: "Prüfungen von Osiris"
-    },
-    5: {
-      modeName: "Schmelztiegel"
-    },
-    10: {
-      modeName: "Kontrolle"
-    },
-    35: {
-      modeName: "Inferno: Vorherrschaft"
-    },
-    33: {
-      modeName: "Rumble: Vorherrschaft"
+      name: 'Vorherrschaft'
     },
     15: {
-      modeName: "Doppel"
+      name: 'Scharlach-Woche'
+    },
+    7: {
+      name: 'PVE'
+    },
+    5: {
+      name: 'Schmelztiegel'
+    },
+    25: {
+      name: 'Hexenkessel'
+    },
+    17: {
+      name: 'Prestige-Dämmerung'
+    },
+    39: {
+      name: 'Prüfungen der Neun'
+    },
+    62: {
+      name: 'Team Versengt'
+    },
+    38: {
+      name: 'Countdown'
+    },
+    12: {
+      name: 'Konflikt'
+    },
+    40: {
+      name: 'Treffpunkt'
+    },
+    61: {
+      name: 'Versengt'
+    },
+    2: {
+      name: 'Story'
+    },
+    49: {
+      name: 'Alle Doppel'
+    },
+    19: {
+      name: 'Eisenbanner'
+    },
+    4: {
+      name: 'Raid'
+    },
+    60: {
+      name: 'Bergung'
+    },
+    37: {
+      name: 'Überleben'
+    },
+    41: {
+      name: 'Prüfungen der Neun: Countdown'
+    },
+    18: {
+      name: 'Strikes'
+    },
+    59: {
+      name: 'Showdown'
+    },
+    57: {
+      name: 'Privatmatches: Rumble'
+    },
+    43: {
+      name: 'Eisenbanner: Kontrolle'
+    },
+    52: {
+      name: 'Privatmatches: Kontrolle'
+    },
+    10: {
+      name: 'Kontrolle'
+    },
+    44: {
+      name: 'Eisenbanner: Konflikt'
+    },
+    6: {
+      name: 'Erkunden'
+    },
+    55: {
+      name: 'Privatmatches: Überleben'
+    },
+    16: {
+      name: 'Dämmerung-Strikes'
+    },
+    50: {
+      name: 'Doppel'
+    },
+    53: {
+      name: 'Privatmatches: Vorherrschaft'
+    },
+    3: {
+      name: 'Normale Strikes'
+    },
+    45: {
+      name: 'Eisenbanner: Vorherrschaft'
     },
   },
   en: {
-    23: {
-      modeName: "Elimination"
-    },
-    25: {
-      modeName: "Mayhem"
-    },
-    6: {
-      modeName: "Patrol"
-    },
-    3: {
-      modeName: "Normal Strikes"
-    },
-    17: {
-      modeName: "Heroic Strikes"
-    },
-    19: {
-      modeName: "Iron Banner"
-    },
-    16: {
-      modeName: "Nightfall Strikes"
+    56: {
+      name: 'Private Matches Mayhem'
     },
     32: {
-      modeName: "Private Matches"
+      name: 'Private Matches'
     },
-    4: {
-      modeName: "Raid"
+    54: {
+      name: 'Private Matches Countdown'
     },
-    8: {
-      modeName: "Introduction"
+    48: {
+      name: 'Rumble'
     },
-    12: {
-      modeName: "Clash"
+    58: {
+      name: 'Heroic Adventure'
     },
-    22: {
-      modeName: "Arena Challenge"
+    71: {
+      name: 'Clash: Quickplay'
     },
-    24: {
-      modeName: "Rift"
+    42: {
+      name: 'Trials of the Nine Survival'
     },
-    36: {
-      modeName: "Mayhem Supremacy"
+    47: {
+      name: 'Scored Prestige Nightfall'
     },
-    27: {
-      modeName: "Mayhem Rumble"
+    46: {
+      name: 'Scored Nightfall Strikes'
+    },
+    51: {
+      name: 'Private Matches Clash'
     },
     0: {
-      modeName: "All"
+      name: 'All'
     },
-    30: {
-      modeName: "Challenge of the Elders"
-    },
-    18: {
-      modeName: "Strikes"
-    },
-    21: {
-      modeName: "Arena"
-    },
-    28: {
-      modeName: "Zone Control"
-    },
-    9: {
-      modeName: "Skirmish"
-    },
-    29: {
-      modeName: "Sparrow Racing"
-    },
-    11: {
-      modeName: "Salvage"
-    },
-    20: {
-      modeName: "Prison of Elders"
-    },
-    34: {
-      modeName: "Supremacy"
-    },
-    13: {
-      modeName: "Rumble"
+    66: {
+      name: 'Forge'
     },
     31: {
-      modeName: "Supremacy"
-    },
-    7: {
-      modeName: "PvE"
-    },
-    26: {
-      modeName: "Mayhem Clash"
-    },
-    2: {
-      modeName: "Story"
-    },
-    14: {
-      modeName: "Trials of Osiris"
-    },
-    5: {
-      modeName: "Crucible"
-    },
-    10: {
-      modeName: "Control"
-    },
-    35: {
-      modeName: "Inferno Supremacy"
-    },
-    33: {
-      modeName: "Rumble Supremacy"
+      name: 'Supremacy'
     },
     15: {
-      modeName: "Doubles"
+      name: 'Crimson Doubles'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Crucible'
+    },
+    25: {
+      name: 'Mayhem'
+    },
+    68: {
+      name: 'Iron Banner Salvage'
+    },
+    74: {
+      name: 'Control: Competitive'
+    },
+    17: {
+      name: 'Prestige Nightfall'
+    },
+    39: {
+      name: 'Trials of the Nine'
+    },
+    62: {
+      name: 'Team Scorched'
+    },
+    75: {
+      name: 'Gambit Prime'
+    },
+    38: {
+      name: 'Countdown'
+    },
+    12: {
+      name: 'Clash'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Scorched'
+    },
+    2: {
+      name: 'Story'
+    },
+    49: {
+      name: 'All Doubles'
+    },
+    19: {
+      name: 'Iron Banner'
+    },
+    63: {
+      name: 'Gambit'
+    },
+    4: {
+      name: 'Raid'
+    },
+    60: {
+      name: 'Salvage'
+    },
+    37: {
+      name: 'Survival'
+    },
+    41: {
+      name: 'Trials of the Nine Countdown'
+    },
+    18: {
+      name: 'Strikes'
+    },
+    64: {
+      name: 'Competitive Co-Op'
+    },
+    69: {
+      name: 'Competitive PvP'
+    },
+    59: {
+      name: 'Showdown'
+    },
+    57: {
+      name: 'Private Matches Rumble'
+    },
+    43: {
+      name: 'Iron Banner Control'
+    },
+    52: {
+      name: 'Private Matches Control'
+    },
+    10: {
+      name: 'Control'
+    },
+    70: {
+      name: 'Quickplay PvP'
+    },
+    44: {
+      name: 'Iron Banner Clash'
+    },
+    6: {
+      name: 'Explore'
+    },
+    73: {
+      name: 'Control: Quickplay'
+    },
+    55: {
+      name: 'Private Matches Survival'
+    },
+    16: {
+      name: 'Nightfall Strikes'
+    },
+    50: {
+      name: 'Doubles'
+    },
+    53: {
+      name: 'Private Matches Supremacy'
+    },
+    76: {
+      name: 'The Reckoning'
+    },
+    67: {
+      name: 'Salvage'
+    },
+    65: {
+      name: 'Breakthrough'
+    },
+    72: {
+      name: 'Clash: Competitive'
+    },
+    3: {
+      name: 'Normal Strikes'
+    },
+    45: {
+      name: 'Iron Banner Supremacy'
     },
   },
   es: {
-    23: {
-      modeName: "Eliminación"
-    },
-    25: {
-      modeName: "Caos"
-    },
-    6: {
-      modeName: "Patrulla"
-    },
-    3: {
-      modeName: "Asaltos normales"
-    },
-    17: {
-      modeName: "Asaltos heroicos"
-    },
-    19: {
-      modeName: "Estandarte de Hierro"
-    },
-    16: {
-      modeName: "Asaltos de Ocaso"
+    56: {
+      name: 'Partidas privadas Caos'
     },
     32: {
-      modeName: "Partidas privadas"
+      name: 'Partidas privadas'
     },
-    4: {
-      modeName: "Incursión"
+    54: {
+      name: 'Partidas privadas Cuenta atrás'
     },
-    8: {
-      modeName: "Introducción"
+    48: {
+      name: 'Disputa'
     },
-    12: {
-      modeName: "Enfrentamiento"
+    58: {
+      name: 'Operación heroica'
     },
-    22: {
-      modeName: "Desafío de la Arena"
+    42: {
+      name: 'Pruebas de los Nueve: Supervivencia'
     },
-    24: {
-      modeName: "Grieta"
+    47: {
+      name: 'Ocaso Prestigio con puntuación'
     },
-    36: {
-      modeName: "Supremacía Caos"
+    46: {
+      name: 'Asaltos Ocaso con puntuación'
     },
-    27: {
-      modeName: "Disputa Caos"
+    51: {
+      name: 'Partidas privadas Enfrentamiento'
     },
     0: {
-      modeName: "Todo"
-    },
-    30: {
-      modeName: "Desafío de los Ancianos"
-    },
-    18: {
-      modeName: "Asaltos"
-    },
-    21: {
-      modeName: "Arena"
-    },
-    28: {
-      modeName: "Control de zonas"
-    },
-    9: {
-      modeName: "Escaramuza"
-    },
-    29: {
-      modeName: "Carreras de Colibríes"
-    },
-    11: {
-      modeName: "Rescate"
-    },
-    20: {
-      modeName: "Presidio de los Ancianos"
-    },
-    34: {
-      modeName: "Supremacía"
-    },
-    13: {
-      modeName: "Disputa"
+      name: 'Todo'
     },
     31: {
-      modeName: "Supremacía"
-    },
-    7: {
-      modeName: "PvE"
-    },
-    26: {
-      modeName: "Enfrentamiento Caos"
-    },
-    2: {
-      modeName: "Historia"
-    },
-    14: {
-      modeName: "Pruebas de Osiris"
-    },
-    5: {
-      modeName: "Crisol"
-    },
-    10: {
-      modeName: "Control"
-    },
-    35: {
-      modeName: "Supremacía infernal"
-    },
-    33: {
-      modeName: "Supremacía Disputa"
+      name: 'Supremacía'
     },
     15: {
-      modeName: "Dobles"
+      name: 'Días Escarlata'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Crisol'
+    },
+    25: {
+      name: 'Caos'
+    },
+    17: {
+      name: 'Ocaso Prestigio'
+    },
+    39: {
+      name: 'Pruebas de los Nueve'
+    },
+    62: {
+      name: 'Equipo calcinado'
+    },
+    38: {
+      name: 'Cuenta atrás'
+    },
+    12: {
+      name: 'Enfrentamiento'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Calcinado'
+    },
+    2: {
+      name: 'Historia'
+    },
+    49: {
+      name: 'Todo dobles'
+    },
+    19: {
+      name: 'Estandarte de Hierro'
+    },
+    4: {
+      name: 'Incursión'
+    },
+    60: {
+      name: 'Rescate'
+    },
+    37: {
+      name: 'Supervivencia'
+    },
+    41: {
+      name: 'Pruebas de los Nueve: Cuenta atrás'
+    },
+    18: {
+      name: 'Asaltos'
+    },
+    59: {
+      name: 'Duelo'
+    },
+    57: {
+      name: 'Partidas privadas Disputa'
+    },
+    43: {
+      name: 'Estandarte de Hierro: Control'
+    },
+    52: {
+      name: 'Partidas privadas Control'
+    },
+    10: {
+      name: 'Control'
+    },
+    44: {
+      name: 'Estandarte de Hierro: Enfrentamiento'
+    },
+    6: {
+      name: 'Explorar'
+    },
+    55: {
+      name: 'Partidas privadas Supervivencia'
+    },
+    16: {
+      name: 'Asaltos de Ocaso'
+    },
+    50: {
+      name: 'Dobles'
+    },
+    53: {
+      name: 'Partidas privadas Supremacía'
+    },
+    3: {
+      name: 'Asaltos normales'
+    },
+    45: {
+      name: 'Estandarte de Hierro: Supremacía'
+    },
+  },
+  esmx: {
+    56: {
+      name: 'Partidas privadas Caos'
+    },
+    32: {
+      name: 'Partidas privadas'
+    },
+    54: {
+      name: 'Partidas privadas Cuenta regresiva'
+    },
+    48: {
+      name: 'Disputa'
+    },
+    58: {
+      name: 'Operación Heroica'
+    },
+    42: {
+      name: 'Pruebas de los Nueve Supervivencia'
+    },
+    47: {
+      name: 'Ocaso Prestigio con puntuación'
+    },
+    46: {
+      name: 'Asaltos Ocaso con puntuación'
+    },
+    51: {
+      name: 'Partidas privadas Enfrentamiento'
+    },
+    0: {
+      name: 'Todo'
+    },
+    31: {
+      name: 'Supremacía'
+    },
+    15: {
+      name: 'Dobles Carmesí.'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Crisol'
+    },
+    25: {
+      name: 'Caos'
+    },
+    17: {
+      name: 'Ocaso Prestigio'
+    },
+    39: {
+      name: 'Pruebas de los Nueve'
+    },
+    62: {
+      name: 'Equipo calcinado'
+    },
+    38: {
+      name: 'Cuenta regresiva'
+    },
+    12: {
+      name: 'Enfrentamiento'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Calcinado'
+    },
+    2: {
+      name: 'Historia'
+    },
+    49: {
+      name: 'Todos los Dobles'
+    },
+    19: {
+      name: 'Estandarte de Hierro'
+    },
+    4: {
+      name: 'Incursión'
+    },
+    60: {
+      name: 'Rescate'
+    },
+    37: {
+      name: 'Supervivencia'
+    },
+    41: {
+      name: 'Pruebas de los Nueve Cuenta regresiva'
+    },
+    18: {
+      name: 'Asaltos'
+    },
+    59: {
+      name: 'Confrontación'
+    },
+    57: {
+      name: 'Partidas privadas Disputa'
+    },
+    43: {
+      name: 'Estandarte de Hierro Control'
+    },
+    52: {
+      name: 'Partidas privadas Control'
+    },
+    10: {
+      name: 'Control'
+    },
+    44: {
+      name: 'Estandarte de Hierro Enfrentamiento'
+    },
+    6: {
+      name: 'Explorar'
+    },
+    55: {
+      name: 'Partidas privadas Supervivencia'
+    },
+    16: {
+      name: 'Asaltos de Ocaso'
+    },
+    50: {
+      name: 'Dobles'
+    },
+    53: {
+      name: 'Partidas privadas Supremacía'
+    },
+    3: {
+      name: 'Asaltos normales'
+    },
+    45: {
+      name: 'Estandarte de Hierro Supremacía '
     },
   },
   fr: {
-    23: {
-      modeName: "Élimination"
-    },
-    25: {
-      modeName: "Pagaille"
-    },
-    6: {
-      modeName: "Patrouille"
-    },
-    3: {
-      modeName: "Assauts normaux"
-    },
-    17: {
-      modeName: "Assauts « Épiques »"
-    },
-    19: {
-      modeName: "Bannière de Fer"
-    },
-    16: {
-      modeName: "Assauts « Nuit noire »"
+    56: {
+      name: 'Parties privées - Pagaille'
     },
     32: {
-      modeName: "Parties privées"
+      name: 'Parties privées'
     },
-    4: {
-      modeName: "Raid"
+    54: {
+      name: 'Parties privées - Explosif'
     },
-    8: {
-      modeName: "Introduction au JcJ"
+    48: {
+      name: 'Carnage'
     },
-    12: {
-      modeName: "Choc"
+    58: {
+      name: 'Aventure Épique'
     },
-    22: {
-      modeName: "Défis de l'Arène"
+    42: {
+      name: 'Jugement des Neuf - Survie'
     },
-    24: {
-      modeName: "Faille"
+    47: {
+      name: 'Nuit noire Prestige avec score'
     },
-    36: {
-      modeName: "Pagaille – Suprématie"
+    46: {
+      name: 'Nuit noire avec score'
     },
-    27: {
-      modeName: "Pagaille – Carnage"
+    51: {
+      name: 'Parties privées - Choc'
     },
     0: {
-      modeName: "Toutes les activités"
-    },
-    30: {
-      modeName: "Défi des vétérans"
-    },
-    18: {
-      modeName: "Assauts"
-    },
-    21: {
-      modeName: "Arène"
-    },
-    28: {
-      modeName: "Contrôle de zone"
-    },
-    9: {
-      modeName: "Escarmouche"
-    },
-    29: {
-      modeName: "Course de Passereaux"
-    },
-    11: {
-      modeName: "Défense"
-    },
-    20: {
-      modeName: "Prison des vétérans"
-    },
-    34: {
-      modeName: "Suprématie"
-    },
-    13: {
-      modeName: "Carnage"
+      name: 'Toutes les activités'
     },
     31: {
-      modeName: "Suprématie"
-    },
-    7: {
-      modeName: "JcE"
-    },
-    26: {
-      modeName: "Pagaille – Choc"
-    },
-    2: {
-      modeName: "Histoire"
-    },
-    14: {
-      modeName: "Jugement d'Osiris"
-    },
-    5: {
-      modeName: "L'Épreuve"
-    },
-    10: {
-      modeName: "Domination"
-    },
-    35: {
-      modeName: "Infernal – Suprématie"
-    },
-    33: {
-      modeName: "Carnage – Suprématie"
+      name: 'Suprématie'
     },
     15: {
-      modeName: "En tandem"
+      name: 'Jours Garance'
+    },
+    7: {
+      name: 'JcE'
+    },
+    5: {
+      name: 'L\'Épreuve'
+    },
+    25: {
+      name: 'Pagaille'
+    },
+    17: {
+      name: 'Nuit noire Prestige'
+    },
+    39: {
+      name: 'Jugement des Neuf'
+    },
+    62: {
+      name: 'Équipe brûlante'
+    },
+    38: {
+      name: 'Explosif'
+    },
+    12: {
+      name: 'Choc'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Brûlant'
+    },
+    2: {
+      name: 'Histoire'
+    },
+    49: {
+      name: 'Tous les modes En tandem'
+    },
+    19: {
+      name: 'Bannière de Fer'
+    },
+    4: {
+      name: 'Raid'
+    },
+    60: {
+      name: 'Défense'
+    },
+    37: {
+      name: 'Survie'
+    },
+    41: {
+      name: 'Jugement des Neuf - Explosif'
+    },
+    18: {
+      name: 'Assauts'
+    },
+    59: {
+      name: 'Confrontation'
+    },
+    57: {
+      name: 'Parties privées - Carnage'
+    },
+    43: {
+      name: 'Bannière de Fer - Domination'
+    },
+    52: {
+      name: 'Parties privées - Domination'
+    },
+    10: {
+      name: 'Domination'
+    },
+    44: {
+      name: 'Bannière de Fer - Choc'
+    },
+    6: {
+      name: 'Exploration'
+    },
+    55: {
+      name: 'Parties privées - Survie'
+    },
+    16: {
+      name: 'Nuit noire'
+    },
+    50: {
+      name: 'En tandem'
+    },
+    53: {
+      name: 'Parties privées - Suprématie'
+    },
+    3: {
+      name: 'Assauts normaux'
+    },
+    45: {
+      name: 'Bannière de Fer - Suprématie'
     },
   },
   it: {
-    23: {
-      modeName: "Eliminazione"
-    },
-    25: {
-      modeName: "Pandemonio"
-    },
-    6: {
-      modeName: "Pattuglia"
-    },
-    3: {
-      modeName: "Assalti normali"
-    },
-    17: {
-      modeName: "Assalti Eroici"
-    },
-    19: {
-      modeName: "Stendardo di Ferro"
-    },
-    16: {
-      modeName: "Assalti Cala la Notte"
+    56: {
+      name: 'Partite private: Pandemonio'
     },
     32: {
-      modeName: "Partite private"
+      name: 'Partite private'
     },
-    4: {
-      modeName: "Incursione"
+    54: {
+      name: 'Partite private: Detonazione'
     },
-    8: {
-      modeName: "Introduzione"
+    48: {
+      name: 'Rissa'
     },
-    12: {
-      modeName: "Scontro"
+    58: {
+      name: 'Avventura eroica'
     },
-    22: {
-      modeName: "Sfide dell'Arena"
+    42: {
+      name: 'Prove dei Nove: Sopravvivenza'
     },
-    24: {
-      modeName: "Frattura"
+    47: {
+      name: 'Cala la Notte autorevole a punti'
     },
-    36: {
-      modeName: "Pandemonio: Supremazia"
+    46: {
+      name: 'Cala la Notte a punti'
     },
-    27: {
-      modeName: "Pandemonio: Rissa"
+    51: {
+      name: 'Partite private: Scontro'
     },
     0: {
-      modeName: "Globale"
-    },
-    30: {
-      modeName: "Sfida degli Anziani"
-    },
-    18: {
-      modeName: "Assalti"
-    },
-    21: {
-      modeName: "Arena"
-    },
-    28: {
-      modeName: "Zone di Controllo"
-    },
-    9: {
-      modeName: "Schermaglia"
-    },
-    29: {
-      modeName: "Gare di Astori"
-    },
-    11: {
-      modeName: "Recupero"
-    },
-    20: {
-      modeName: "Prigione degli Anziani"
-    },
-    34: {
-      modeName: "Supremazia"
-    },
-    13: {
-      modeName: "Rissa"
+      name: 'Tutte'
     },
     31: {
-      modeName: "Supremazia"
-    },
-    7: {
-      modeName: "PvE"
-    },
-    26: {
-      modeName: "Pandemonio: Scontro"
-    },
-    2: {
-      modeName: "Storia"
-    },
-    14: {
-      modeName: "Prove di Osiride"
-    },
-    5: {
-      modeName: "Crogiolo"
-    },
-    10: {
-      modeName: "Controllo"
-    },
-    35: {
-      modeName: "Inferno: Supremazia"
-    },
-    33: {
-      modeName: "Supremazia: Rissa"
+      name: 'Supremazia'
     },
     15: {
-      modeName: "Doppietta"
+      name: 'Giorni Scarlatti'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Crogiolo'
+    },
+    25: {
+      name: 'Pandemonio'
+    },
+    17: {
+      name: 'Cala la Notte autorevole'
+    },
+    39: {
+      name: 'Prove dei Nove'
+    },
+    62: {
+      name: 'Squadra ardente'
+    },
+    38: {
+      name: 'Detonazione'
+    },
+    12: {
+      name: 'Scontro'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Ardente'
+    },
+    2: {
+      name: 'Storia'
+    },
+    49: {
+      name: 'Tutte le Doppiette'
+    },
+    19: {
+      name: 'Stendardo di Ferro'
+    },
+    4: {
+      name: 'Incursione'
+    },
+    60: {
+      name: 'Recupero'
+    },
+    37: {
+      name: 'Sopravvivenza'
+    },
+    41: {
+      name: 'Prove dei Nove: Detonazione'
+    },
+    18: {
+      name: 'Assalti'
+    },
+    59: {
+      name: 'Prova di forza'
+    },
+    57: {
+      name: 'Partite private: Rissa'
+    },
+    43: {
+      name: 'Stendardo di Ferro: Controllo'
+    },
+    52: {
+      name: 'Partite private: Controllo'
+    },
+    10: {
+      name: 'Controllo'
+    },
+    44: {
+      name: 'Stendardo di Ferro: Scontro'
+    },
+    6: {
+      name: 'Esplora'
+    },
+    55: {
+      name: 'Partite private: Sopravvivenza'
+    },
+    16: {
+      name: 'Assalti Cala la Notte'
+    },
+    50: {
+      name: 'Doppietta'
+    },
+    53: {
+      name: 'Partite private: Supremazia'
+    },
+    3: {
+      name: 'Assalti normali'
+    },
+    45: {
+      name: 'Stendardo di Ferro: Supremazia'
     },
   },
   ja: {
-    23: {
-      modeName: "サバイバル"
-    },
-    25: {
-      modeName: "カオス"
-    },
-    6: {
-      modeName: "パトロール"
-    },
-    3: {
-      modeName: "普通のストライク"
-    },
-    17: {
-      modeName: "英雄ストライク"
-    },
-    19: {
-      modeName: "アイアンバナー"
-    },
-    16: {
-      modeName: "ナイトフォールストライク"
+    56: {
+      name: 'プライベート対戦(カオス)'
     },
     32: {
-      modeName: "非公開対戦"
+      name: '非公開対戦'
     },
-    4: {
-      modeName: "レイド"
+    54: {
+      name: 'プライベート対戦(カウントダウン)'
     },
-    8: {
-      modeName: "イントロ"
+    48: {
+      name: 'ランブル'
     },
-    12: {
-      modeName: "クラッシュ"
+    58: {
+      name: '英雄アドベンチャー'
     },
-    22: {
-      modeName: "アリーナの挑戦"
+    42: {
+      name: 'ナインの試練「ライフリミット」'
     },
-    24: {
-      modeName: "暁光"
+    47: {
+      name: 'スコア付きナイトフォール(威光)'
     },
-    36: {
-      modeName: "カオス: 闘魂"
+    46: {
+      name: 'スコア付きナイトフォールストライク'
     },
-    27: {
-      modeName: "カオス: ランブル"
+    51: {
+      name: 'プライベート対戦(クラッシュ)'
     },
     0: {
-      modeName: "全て"
-    },
-    30: {
-      modeName: "エルダーの挑戦"
-    },
-    18: {
-      modeName: "ストライク"
-    },
-    21: {
-      modeName: "アリーナ"
-    },
-    28: {
-      modeName: "エリアコントロール"
-    },
-    9: {
-      modeName: "スカーミッシュ"
-    },
-    29: {
-      modeName: "スパローレース"
-    },
-    11: {
-      modeName: "サルベージ"
-    },
-    20: {
-      modeName: "エルダーズ・プリズン"
-    },
-    34: {
-      modeName: "闘魂"
-    },
-    13: {
-      modeName: "ランブル"
+      name: '全て'
     },
     31: {
-      modeName: "闘魂"
-    },
-    7: {
-      modeName: "PvE"
-    },
-    26: {
-      modeName: "カオス: クラッシュ"
-    },
-    2: {
-      modeName: "ストーリー"
-    },
-    14: {
-      modeName: "オシリスの試練"
-    },
-    5: {
-      modeName: "クルーシブル"
-    },
-    10: {
-      modeName: "コントロール"
-    },
-    35: {
-      modeName: "インフェルノ: 闘魂"
-    },
-    33: {
-      modeName: "ランブル: 闘魂"
+      name: '闘魂'
     },
     15: {
-      modeName: "ダブル"
+      name: '真紅の日々'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'クルーシブル'
+    },
+    25: {
+      name: 'カオス'
+    },
+    17: {
+      name: 'ナイトフォール(威光)'
+    },
+    39: {
+      name: 'ナインの試練'
+    },
+    62: {
+      name: '焼け焦げチーム'
+    },
+    38: {
+      name: 'カウントダウン'
+    },
+    12: {
+      name: 'クラッシュ'
+    },
+    40: {
+      name: 'ソーシャル'
+    },
+    61: {
+      name: '焼け焦げ'
+    },
+    2: {
+      name: 'ストーリー'
+    },
+    49: {
+      name: '全ダブル'
+    },
+    19: {
+      name: 'アイアンバナー'
+    },
+    4: {
+      name: 'レイド'
+    },
+    60: {
+      name: 'サルベージ'
+    },
+    37: {
+      name: 'ライフリミット'
+    },
+    41: {
+      name: 'ナインの試練「カウントダウン」'
+    },
+    18: {
+      name: 'ストライク'
+    },
+    59: {
+      name: 'ショーダウン'
+    },
+    57: {
+      name: 'プライベート対戦(ランブル)'
+    },
+    43: {
+      name: 'アイアンバナー「コントロール」'
+    },
+    52: {
+      name: 'プライベート対戦(コントロール)'
+    },
+    10: {
+      name: 'コントロール'
+    },
+    44: {
+      name: 'アイアンバナー「クラッシュ」'
+    },
+    6: {
+      name: '探索'
+    },
+    55: {
+      name: 'プライベート対戦(ライフリミット)'
+    },
+    16: {
+      name: 'ナイトフォールストライク'
+    },
+    50: {
+      name: 'ダブル'
+    },
+    53: {
+      name: 'プライベート対戦(闘魂)'
+    },
+    3: {
+      name: '普通のストライク'
+    },
+    45: {
+      name: 'アイアンバナー「闘魂」'
+    },
+  },
+  pl: {
+    56: {
+      name: 'Mecze prywatne – Jatka'
+    },
+    32: {
+      name: 'Mecze prywatne'
+    },
+    54: {
+      name: 'Mecze prywatne – Odliczanie'
+    },
+    48: {
+      name: 'Jazgot'
+    },
+    58: {
+      name: 'Heroiczna przygoda'
+    },
+    42: {
+      name: 'Przetrwanie – Próby Dziewiątki'
+    },
+    47: {
+      name: 'Punktowany prestiżowy nocny szturm'
+    },
+    46: {
+      name: 'Punktowane nocne szturmy'
+    },
+    51: {
+      name: 'Mecze prywatne – Starcie'
+    },
+    0: {
+      name: 'Wszystko'
+    },
+    31: {
+      name: 'Supremacja'
+    },
+    15: {
+      name: 'Karmazynowe Dni'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Tygiel'
+    },
+    25: {
+      name: 'Jatka'
+    },
+    17: {
+      name: 'Prestiżowy nocny szturm'
+    },
+    39: {
+      name: 'Próby Dziewiątki'
+    },
+    62: {
+      name: 'Spopielona drużyna'
+    },
+    38: {
+      name: 'Odliczanie'
+    },
+    12: {
+      name: 'Starcie'
+    },
+    40: {
+      name: 'Społecznościowe'
+    },
+    61: {
+      name: 'Spopielenie'
+    },
+    2: {
+      name: 'Fabuła'
+    },
+    49: {
+      name: 'Wszystkie Deble'
+    },
+    19: {
+      name: 'Żelazna Chorągiew'
+    },
+    4: {
+      name: 'Najazd'
+    },
+    60: {
+      name: 'Odzyskanie'
+    },
+    37: {
+      name: 'Przetrwanie'
+    },
+    41: {
+      name: 'Próby Dziewiątki – Odliczanie'
+    },
+    18: {
+      name: 'Szturmy'
+    },
+    59: {
+      name: 'Rozstrzygnięcie'
+    },
+    57: {
+      name: 'Mecze prywatne – Jazgot'
+    },
+    43: {
+      name: 'Żelazna Chorągiew – Kontrola'
+    },
+    52: {
+      name: 'Mecze prywatne – Kontrola'
+    },
+    10: {
+      name: 'Kontrola'
+    },
+    44: {
+      name: 'Starcie – Żelazna Chorągiew'
+    },
+    6: {
+      name: 'Eksploracja'
+    },
+    55: {
+      name: 'Mecze prywatne – Przetrwanie'
+    },
+    16: {
+      name: 'Nocne szturmy'
+    },
+    50: {
+      name: 'Debel'
+    },
+    53: {
+      name: 'Mecze prywatne – Supremacja'
+    },
+    3: {
+      name: 'Zwykłe szturmy'
+    },
+    45: {
+      name: 'Supremacja – Żelazna Chorągiew'
     },
   },
   ptbr: {
-    23: {
-      modeName: "Eliminação"
-    },
-    25: {
-      modeName: "Caos"
-    },
-    6: {
-      modeName: "Patrulha"
-    },
-    3: {
-      modeName: "Assaltos normais"
-    },
-    17: {
-      modeName: "Assaltos Heroicos"
-    },
-    19: {
-      modeName: "Bandeira de Ferro"
-    },
-    16: {
-      modeName: "Assaltos do Anoitecer"
+    56: {
+      name: 'Partidas privadas – Caos'
     },
     32: {
-      modeName: "Partidas privadas"
+      name: 'Partidas privadas'
     },
-    4: {
-      modeName: "Incursão"
+    54: {
+      name: 'Partidas privadas – Detonação'
     },
-    8: {
-      modeName: "Introdução"
+    48: {
+      name: 'Briga'
     },
-    12: {
-      modeName: "Enfrentamento"
+    58: {
+      name: 'Aventura Heroica'
     },
-    22: {
-      modeName: "Desafio da Arena"
+    42: {
+      name: 'Desafios dos Nove – Sobrevivência'
     },
-    24: {
-      modeName: "Fissura"
+    47: {
+      name: 'Anoitecer Prestígio Pontuado'
     },
-    36: {
-      modeName: "Caos: Supremacia"
+    46: {
+      name: 'Assaltos do Anoitecer Pontuados'
     },
-    27: {
-      modeName: "Caos: Briga"
+    51: {
+      name: 'Partidas privadas – Enfrentamento'
     },
     0: {
-      modeName: "Tudo"
-    },
-    30: {
-      modeName: "Desafio dos Anciões"
-    },
-    18: {
-      modeName: "Assaltos"
-    },
-    21: {
-      modeName: "Arena"
-    },
-    28: {
-      modeName: "Controle de zonas"
-    },
-    9: {
-      modeName: "Disputa"
-    },
-    29: {
-      modeName: "Corridas de Pardais"
-    },
-    11: {
-      modeName: "Recuperação"
-    },
-    20: {
-      modeName: "Prisão dos Anciões"
-    },
-    34: {
-      modeName: "Supremacia"
-    },
-    13: {
-      modeName: "Briga"
+      name: 'Tudo'
     },
     31: {
-      modeName: "Supremacia"
-    },
-    7: {
-      modeName: "PvE"
-    },
-    26: {
-      modeName: "Caos: Enfrentamento"
-    },
-    2: {
-      modeName: "História"
-    },
-    14: {
-      modeName: "Desafios de Osíris"
-    },
-    5: {
-      modeName: "Crisol"
-    },
-    10: {
-      modeName: "Controle"
-    },
-    35: {
-      modeName: "Inferno: Supremacia"
-    },
-    33: {
-      modeName: "Briga: Supremacia"
+      name: 'Supremacia'
     },
     15: {
-      modeName: "Duplas"
+      name: 'Dias Carmesins'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Crisol'
+    },
+    25: {
+      name: 'Caos'
+    },
+    17: {
+      name: 'Anoitecer Prestígio'
+    },
+    39: {
+      name: 'Desafios dos Nove'
+    },
+    62: {
+      name: 'Team Scorched'
+    },
+    38: {
+      name: 'Detonação'
+    },
+    12: {
+      name: 'Enfrentamento'
+    },
+    40: {
+      name: 'Social'
+    },
+    61: {
+      name: 'Scorched'
+    },
+    2: {
+      name: 'História'
+    },
+    49: {
+      name: 'Todas Duplas'
+    },
+    19: {
+      name: 'Bandeira de Ferro'
+    },
+    4: {
+      name: 'Incursão'
+    },
+    60: {
+      name: 'Recuperação'
+    },
+    37: {
+      name: 'Sobrevivência'
+    },
+    41: {
+      name: 'Desafios dos Nove – Detonação'
+    },
+    18: {
+      name: 'Assaltos'
+    },
+    59: {
+      name: 'Confronto'
+    },
+    57: {
+      name: 'Partidas privadas – Briga'
+    },
+    43: {
+      name: 'Bandeira de Ferro – Controle'
+    },
+    52: {
+      name: 'Partidas privadas – Controle'
+    },
+    10: {
+      name: 'Controle'
+    },
+    44: {
+      name: 'Bandeira de Ferro – Enfrentamento'
+    },
+    6: {
+      name: 'Explore'
+    },
+    55: {
+      name: 'Partidas privadas – Sobrevivência'
+    },
+    16: {
+      name: 'Assaltos do Anoitecer'
+    },
+    50: {
+      name: 'Duplas'
+    },
+    53: {
+      name: 'Partidas privadas – Supremacia'
+    },
+    3: {
+      name: 'Assaltos normais'
+    },
+    45: {
+      name: 'Bandeira de Ferro – Supremacia'
+    },
+  },
+  ru: {
+    56: {
+      name: 'Закрытые матчи: "Хаос"'
+    },
+    32: {
+      name: 'Закрытые матчи'
+    },
+    54: {
+      name: 'Закрытые матчи: "Обратный отсчет"'
+    },
+    48: {
+      name: 'Стычка'
+    },
+    58: {
+      name: 'Героическое приключение'
+    },
+    42: {
+      name: 'Испытания Девяти: Выживание'
+    },
+    47: {
+      name: 'Сумрачные налеты "Престиж" с подсчетом очков'
+    },
+    46: {
+      name: 'Сумрачные налеты с подсчетом очков'
+    },
+    51: {
+      name: 'Закрытые матчи: "Столкновение"'
+    },
+    0: {
+      name: 'Все'
+    },
+    31: {
+      name: 'Превосходство'
+    },
+    15: {
+      name: 'Багряные дни '
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: 'Горнило'
+    },
+    25: {
+      name: 'Хаос'
+    },
+    17: {
+      name: 'Сумрачные налеты "Престиж"'
+    },
+    39: {
+      name: 'Испытания Девяти'
+    },
+    62: {
+      name: 'Команда "Ожог"'
+    },
+    38: {
+      name: 'Обратный отсчет'
+    },
+    12: {
+      name: 'Столкновение'
+    },
+    40: {
+      name: 'Общение'
+    },
+    61: {
+      name: 'Ожог'
+    },
+    2: {
+      name: 'Сюжет'
+    },
+    49: {
+      name: 'Все "Напарники"'
+    },
+    19: {
+      name: 'Железное знамя'
+    },
+    4: {
+      name: 'Рейды'
+    },
+    60: {
+      name: 'Спасение'
+    },
+    37: {
+      name: 'Выживание'
+    },
+    41: {
+      name: 'Испытания Девяти: Обратный отсчет'
+    },
+    18: {
+      name: 'Налеты'
+    },
+    59: {
+      name: 'Поединок'
+    },
+    57: {
+      name: 'Закрытые матчи: "Стычка"'
+    },
+    43: {
+      name: 'Железное знамя: Контроль'
+    },
+    52: {
+      name: 'Закрытые матчи: "Контроль"'
+    },
+    10: {
+      name: 'Контроль'
+    },
+    44: {
+      name: 'Железное знамя: Столкновение'
+    },
+    6: {
+      name: 'Режим исследования'
+    },
+    55: {
+      name: 'Закрытые матчи: "Выживание"'
+    },
+    16: {
+      name: 'Сумрачные налеты'
+    },
+    50: {
+      name: 'Напарники'
+    },
+    53: {
+      name: 'Закрытые матчи: "Превосходство"'
+    },
+    3: {
+      name: 'Обычные налеты'
+    },
+    45: {
+      name: 'Железное знамя: Превосходство'
+    },
+  },
+  zhcht: {
+    56: {
+      name: '私人對戰激戰模式'
+    },
+    32: {
+      name: '私人對戰'
+    },
+    54: {
+      name: '私人對戰分秒必爭'
+    },
+    48: {
+      name: '混戰模式'
+    },
+    58: {
+      name: '英雄探險任務'
+    },
+    42: {
+      name: '九巨頭試煉生存模式'
+    },
+    47: {
+      name: '計分榮耀日暮任務'
+    },
+    46: {
+      name: '計分日暮突擊任務'
+    },
+    51: {
+      name: '私人對戰死鬥模式'
+    },
+    0: {
+      name: '全部'
+    },
+    31: {
+      name: '奪寶模式'
+    },
+    15: {
+      name: '腥紅雙打'
+    },
+    7: {
+      name: 'PvE'
+    },
+    5: {
+      name: '熔爐競技場'
+    },
+    25: {
+      name: '激戰模式'
+    },
+    17: {
+      name: '榮耀日暮任務'
+    },
+    39: {
+      name: '九巨頭試煉'
+    },
+    62: {
+      name: 'Team Scorched'
+    },
+    38: {
+      name: '分秒必爭'
+    },
+    12: {
+      name: '死鬥模式'
+    },
+    40: {
+      name: '社交'
+    },
+    61: {
+      name: 'Scorched'
+    },
+    2: {
+      name: '主線'
+    },
+    49: {
+      name: '所有雙打'
+    },
+    19: {
+      name: '鋼鐵旗幟'
+    },
+    4: {
+      name: '掠奪任務'
+    },
+    60: {
+      name: '搶救模式'
+    },
+    37: {
+      name: '生存模式'
+    },
+    41: {
+      name: '九巨頭試煉分秒必爭模式'
+    },
+    18: {
+      name: '突擊任務'
+    },
+    59: {
+      name: '正面對決'
+    },
+    57: {
+      name: '私人對戰混戰模式'
+    },
+    43: {
+      name: '鋼鐵旗幟攻佔模式'
+    },
+    52: {
+      name: '私人對戰攻佔模式'
+    },
+    10: {
+      name: '攻佔模式'
+    },
+    44: {
+      name: '鋼鐵旗幟死鬥模式'
+    },
+    6: {
+      name: '探索'
+    },
+    55: {
+      name: '私人對戰生存模式'
+    },
+    16: {
+      name: '日暮突擊任務'
+    },
+    50: {
+      name: '雙打'
+    },
+    53: {
+      name: '私人對戰奪寶模式'
+    },
+    3: {
+      name: '普通突擊任務'
+    },
+    45: {
+      name: '鋼鐵旗幟奪寶模式'
+    },
+  },
+  data: {
+    116827562: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_52f64830161b33289d36edc03f6d2ea6.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    122250361: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    152599670: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    157639802: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_5b371fef4ecafe733ad487a8fae3b9f5.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    175275639: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_abf5e799b91b9a4dea7f8a7e00106d41.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_story_1.jpg',
+      category: '1'
+    },
+    244324567: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    470484296: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f4b12d402bb36270cd911f4e1fa8bb0c.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    532484583: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_abf5e799b91b9a4dea7f8a7e00106d41.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    547513715: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_234e7e18549d5eae2ddb012f2bcb203a.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    575409284: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_cf348b63307220263eff699dc56bd081.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    748895195: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f6de6d95f600f199c9a674c73cbefbcc.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_allmodes.jpg',
+      category: '0'
+    },
+    803838459: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_d08435f6d51f7ff593c92e7f1394e933.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    910991990: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_ca750af7ed2f0ec3c8e2f019c85f7bab.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    946648766: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f9d1438e0536f36e868e691eb0acfc4d.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1164760493: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_5f8a923a0d0ac1e4289ae3be03f94aa2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_story_1.jpg',
+      category: '1'
+    },
+    1164760504: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1264443021: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1317961215: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fe57052d7cf971f7502daa75a2ca2437.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1340118533: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_922ae83203f6134bae00937d12eab3a2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1350109474: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_abf5e799b91b9a4dea7f8a7e00106d41.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    1370326378: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f4b12d402bb36270cd911f4e1fa8bb0c.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1372032677: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1418469392: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_1e392d102336226d85cc12bc5dbe1ff0.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '3'
+    },
+    1505888634: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1585787840: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1589650888: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f6de6d95f600f199c9a674c73cbefbcc.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '0'
+    },
+    1668923154: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1686739444: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_5f8a923a0d0ac1e4289ae3be03f94aa2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_story_1.jpg',
+      category: '1'
+    },
+    1703282980: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f9d1438e0536f36e868e691eb0acfc4d.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1826469369: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fe57052d7cf971f7502daa75a2ca2437.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    1848252830: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_96f7e9009d4f26e30cfd60564021925e.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '3'
+    },
+    2043403989: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_bfe80e3dafe6686a9dc42df0606bdc9b.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_raid_1.jpg',
+      category: '1'
+    },
+    2096553452: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2239249083: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2263766091: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f4b12d402bb36270cd911f4e1fa8bb0c.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2394616003: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_38e26baf417d26bb3548d97bf4872b54.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    2449176354: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_5f8a923a0d0ac1e4289ae3be03f94aa2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_story_1.jpg',
+      category: '3'
+    },
+    2486723318: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2546784302: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2553544158: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_5b371fef4ecafe733ad487a8fae3b9f5.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    2843684868: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fe57052d7cf971f7502daa75a2ca2437.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3024701316: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_922ae83203f6134bae00937d12eab3a2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3199098480: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_922ae83203f6134bae00937d12eab3a2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3425110680: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3479216388: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fe57052d7cf971f7502daa75a2ca2437.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3497767639: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_0aa1d7b0e0ac2c6820036b6b3dde3e5b.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_patrol_1.jpg',
+      category: '1'
+    },
+    3673334631: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_922ae83203f6134bae00937d12eab3a2.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3780759439: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3789021730: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_234e7e18549d5eae2ddb012f2bcb203a.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    3821502017: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_f9d1438e0536f36e868e691eb0acfc4d.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3862044762: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_ca750af7ed2f0ec3c8e2f019c85f7bab.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    3894474826: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_e74b3385c5269da226372df8ae7f500d.png',
+      pgcrImage: '/img/destiny_content/pgcr/reckoning.jpg',
+      category: '1'
+    },
+    3904624892: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    4033000329: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    4052075509: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fb3e9149c43f7a2e8f8b66cbea7845fe.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
+    },
+    4110605575: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_38e26baf417d26bb3548d97bf4872b54.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_strikes_1.jpg',
+      category: '1'
+    },
+    4212882650: {
+      icon: '/common/destiny2_content/icons/DestinyActivityModeDefinition_fe57052d7cf971f7502daa75a2ca2437.png',
+      pgcrImage: '/img/theme/destiny/bgs/stats/banner_crucible_1.jpg',
+      category: '2'
     },
   },
 };
